refactor(producer): extract validation to middleware
- Create validateComplaintRequest middleware (SRP §3.1)
- Remove validation logic from complaints.service.ts
- Register middleware in complaints.routes.ts before controller
- Service now focuses solely on ticket construction and publishing
