+++ backend/producer/src/routes/complaints.routes.ts
@@ -1,9 +1,10 @@
 import { Router } from 'express';
 import { complaintsController } from '../controllers/complaints.controller.js';
+import { validateComplaintRequest } from '../middlewares/validateComplaintRequest.js';
 
 const router = Router();
 
-// POST /complaints - Create a new complaint (publish event)
-router.post('/', complaintsController.createComplaint);
+// POST /complaints - Validate input, then create a new complaint (publish event)
+router.post('/', validateComplaintRequest, complaintsController.createComplaint);
 
 export { router as complaintsRouter };
